+++ macro/agnostic/generic.sh
@@ -43,6 +43,18 @@ ppl--generic() {
         "NPM") ppl--npm PUBLISH "$@";;
         "ENP") ppl--enp PUBLISH "$@";;
       esac;;
+    "PUBLISH-IMAGE")
+      case "$project_type" in
+        "MVN") _FATAL "NOT IMPLEMENTED";;
+        "NPM") _FATAL "NOT IMPLEMENTED";;
+        "ENP") ppl--enp PUBLISH-IMAGE "$@";;
+      esac;;
+    "SCAN-IMAGE")
+      case "$project_type" in
+        "MVN") _FATAL "NOT IMPLEMENTED";;
+        "NPM") _FATAL "NOT IMPLEMENTED";;
+        "ENP") ppl--enp SCAN-IMAGE "$@";;
+      esac;;
     "POST-DEP-TESTS")
       case "$project_type" in
         "MVN") ppl--mvn POST-DEP-TESTS "$@";;
@@ -63,6 +75,7 @@ ppl--generic() {
       case "$project_type" in
         "MVN") ppl--mvn.generate-build-cache-key "$VARIABLE_NAME";;
         "NPM") _FATAL "Not implemented";;
+        "ENP") ppl--enp.generate-build-cache-key "$VARIABLE_NAME";;
       esac;;
     "GENERATE-BUILD-TARGET-DIR")
       START_SIMPLE_MACRO "$action" "$@"
@@ -72,7 +85,7 @@ ppl--generic() {
       case "$project_type" in
         "MVN") echo "$VARIABLE_NAME=target";;
         "NPM") _FATAL "Not implemented";;
-        "ENP") echo "$VARIABLE_NAME=build";;
+        "ENP") ppl--enp.generate-build-dir-path "$VARIABLE_NAME";;
       esac;;
     *)
       _FATAL "Invalid macro action \"$action\""